requirements.txt
altair==6.0.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
curl_cffi==0.13.0
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
loguru==0.7.3
MarkupSafe==3.0.3
multitasking==0.0.12
narwhals==2.18.0
newsapi-python==0.2.7
nltk==3.9.3
numpy==2.4.3
packaging==26.0
pandas==2.3.3
peewee==4.0.2
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
protobuf==6.33.6
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
streamlit==1.55.0
tenacity==9.1.4
textblob==0.19.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vaderSentiment==3.3.2
watchdog==6.0.0
websockets==16.0
win32_setctime==1.2.0
yfinance==1.2.0
